+++ certbot-apache/certbot_apache/_internal/override_centos.py
@@ -13,7 +13,7 @@ from acme.magic_typing import List  # pylint: disable=unused-import, no-name-in-
 
 from certbot_apache._internal import apache_util
 from certbot_apache import configurator
-from certbot_apache import parser
+from certbot_apache._internal import parser
 
 
 logger = logging.getLogger(__name__)
