+++ certbot-apache/certbot_apache/_internal/override_fedora.py
@@ -8,7 +8,7 @@ from certbot import util
 
 from certbot_apache._internal import apache_util
 from certbot_apache import configurator
-from certbot_apache import parser
+from certbot_apache._internal import parser
 
 
 @zope.interface.provider(interfaces.IPluginFactory)
