+++ certbot-apache/certbot_apache/_internal/override_gentoo.py
@@ -7,7 +7,7 @@ from certbot import interfaces
 
 from certbot_apache._internal import apache_util
 from certbot_apache import configurator
-from certbot_apache import parser
+from certbot_apache._internal import parser
 
 @zope.interface.provider(interfaces.IPluginFactory)
 class GentooConfigurator(configurator.ApacheConfigurator):
