+++ certbot-apache/certbot_apache/http_01.py
@@ -9,7 +9,7 @@ from certbot.compat import filesystem
 from certbot.plugins import common
 
 from certbot_apache.obj import VirtualHost  # pylint: disable=unused-import
-from certbot_apache.parser import get_aug_path
+from certbot_apache._internal.parser import get_aug_path
 
 logger = logging.getLogger(__name__)
