+++ certbot-apache/certbot_apache/tests/centos6_test.py
@@ -6,7 +6,7 @@ from certbot.errors import MisconfigurationError
 
 from certbot_apache import obj
 from certbot_apache._internal import override_centos
-from certbot_apache import parser
+from certbot_apache._internal import parser
 from certbot_apache.tests import util
